9/29 v0.7 2pick and new zone: Shark

+++ Assets/Scripts/Game/Battle/BattlePlace.cs
@@ -12,7 +12,7 @@ public class BattlePlace
     public int Count => cards.Count;
     public int AvailableCount => MaxCount - Count;
     public bool IsFull => Count >= MaxCount;
-    public Dictionary<string, float> options = new Dictionary<string, float>();
+    public Dictionary<string, int> options = new Dictionary<string, int>();
 
     public BattlePlace() {}
     public BattlePlace(List<BattleCard> battleCards) {
@@ -23,7 +23,7 @@ public class BattlePlace
     public BattlePlace(BattlePlace rhs) {
         cards = rhs.cards.Select(x => (x == null) ? null : new BattleCard(x)).ToList();
         MaxCount = rhs.MaxCount;
-        options = new Dictionary<string, float>(rhs.options);
+        options = new Dictionary<string, int>(rhs.options);
     }
 
     public virtual BattlePlaceId GetPlaceId() {
@@ -34,7 +34,7 @@ public class BattlePlace
         return cards.Contains(battleCard);
     }
 
-    public virtual float GetIdentifier(string id) {
+    public virtual int GetIdentifier(string id) {
         var trimId = id.Split('.')[0];
 
         if (id.StartsWith("[")) {
@@ -60,11 +60,11 @@ public class BattlePlace
         };
     }
 
-    public virtual void SetIdentifier(string id, float num) {
+    public virtual void SetIdentifier(string id, int num) {
         options.Set(id, num);
     }
 
-    public virtual void AddIdentifier(string id, float num) {
+    public virtual void AddIdentifier(string id, int num) {
         SetIdentifier(id, GetIdentifier(id) + num);
     }
 }
